Changed name/functionality of class
Changed name to something more meaningful to its purpose.

Functionality is mostly the same, it is now a static class with the
purpose only to convert the incoming TCP stream data and pass it to the
BodyPositionManager.

Added movement tracking. person model now moves to mirror the
transformation movement of the user.

Removed unnecessary comments and debugs